Add files via upload
Binary Search is a search algorithm that works by repeatedly dividing the search interval in half. It is used to search for a specific element in a sorted list or array, and has a time complexity of O(log n).

Binary Search Tree, on the other hand, is a binary tree data structure in which each node has at most two children, with left child being smaller than its parent and right child being greater. This allows for efficient searching, insertion and deletion of elements in a sorted manner, with a worst-case time complexity of O(n) and average-case time complexity of O(log n).